wrapperutils: Use native C++ allocation and deletion for wrappers
When allocating a wrapped object we've some unnecessarily complicated
initializiation logic, where we initially allocate the memory (zero-ing
it) and then calling the constructor for the allocated memory.
The same process is repeatead (in inverted order) when deallocating it.

While this may have the benefit of zero-ing the memory for us, it's just
unncessary as C++ constructors and destructors can handle this for us,
and we can avoid duplicate writes on the same memory space, leaving the
initialization fully to the constructors.

So, using simple new/delete and adjusting the few places where we didn't
initialize the private flags at construction.