arg: Fix MIN/MAX safe big integer limits
Current JS Number's MAX_SAFE_INTEGER value is set to 9007199254740991,
however we set this value to 9007199254740992.

Fix the number computation and add an unit test to ensure that the value
we use is always matching JS definition (so that we are protected in the
unlikely case that anything would change).