function: Early-initialize arguments cache or we won't ever release them
If we throw during the arguments cache creation as we fail during an early
stage of init_cached_function_data, we don't save the function GICallableInfo
or the already allocated parameters, but we'd just return loosing
track of allocated data.

This is not just a potential leak, but also wrong as if a method
introspection is malformed or not fully supported by gjs, we'd just
assert crashing badly during function unallocation because no function
info is set, instead of just throwing as expected.

As per this, let's save the function private informations (such as the
callable info and the arguments cache) as soon as we allocate them,
and update the number of initialized arguments in real time, so that
we can partially de-allocate the arguments cache in case of errors.
This also implies that during function de-init we must ensure that
we don't overflow using the marshallers pointer to check if we've ever
initialized such argument cache and to figure out how many we've allocated
other than of the total number of the introspected function arguments.