gjs-tests: Add gjs_arg_* tests
Ensure that gjs_arg_set, gjs_arg_get and gjs_arg_unset work as expected
using random numbers that are generated per type and using some template
functions to do smarter comparisions.

Expose the random sed we used to generate the random numbers and
generate the c++ one using a random number generated from g_test so that
it will be consistent with `--seed` or in general with the GLib seed.

Add support for a `--cpp-seed` option to be able to pass it one manually