profiler: Support external SysprofCaptureWriters
Using environment variables to control the profiler is limiting,
and when it comes to GNOME Shell, it adds an extra level of
complexity, since it forces us to start GNOME Shell with the
right environment set.

Add support for setting a caller-controlled SysprofCaptureWriter.
The header file treats it as a gpointer to avoid including config.h.
If an external capture writer is set, it takes precedence over the
filename and the file descriptor.