jsapi-util: Add copy and equality operator to GjsAutoPointer
Not being an unique_ptr anymore we've some more freedom (that will be
way more when C++ will allow us to do more static checks), one of the
good gain we've now, is that we can easily implement a copy and swap
idiom to allow to pass the wrappers by value ensuring an internal copy.

This will not work (unfortunately only with runtime checks! :-( ) if the
wrapper doesn't provide a copy function, but this is still nice.

Tests included.