gi: Use GjsSmartPointer to handle destruction of Wrapped objects
With the introduction of GjsSmartPointer we can easily generate smart
pointer types at build-time depending on template arguments and with the
same size of a normal pointer, so use this for WrapperUtils.

Don't change the default behavior for base class, so introduce a fake
smart-pointer using a custom type that has no free function set, while
use it to handle automatically some basic types we have (so far just
GError and GObject will take advantage).

However this opens the gates to a creation of custom specializations of
BoxedInstance that we may use to create different compile-time codepaths
avoiding various dynamic checks.