arg: Use only one single template-based function to create simple types array
When creating arrays for simple types we can just reuse a single
function with typed parameters and reuse functions to fetch the value
from JS, storing it in the proper container.

Added a specialization for boolean types and GType. However, since we need
to make sure it's not confused with the same typedef'ed original type, we
need to use the type tag as well.

As per this, convert the js-value-inl utility functions to template
functions with implicit values, and specialize the ones we need to.

In particular we need a js_value_to_c() specialization for the GType tag
that gets an object and fetches the gtype from it.
We can just return in case of errors, as the generic function will take
care of throwing, in case.

All this permits a nice code cleanup and more type-safety.

(Philip: Added char32_t specializations)