arg: Use argument flags for the gjs conversion functions
Make arguments more readable and easier to extend when it comes to pass
flags to such functions, instead of rely on hard-to-maintain booleans.

Using an enum class that needs some more redefinitions but gives more
type safety, as per this adding a enum-utils.h header with some smarter
templated definitions for operators that match all the enum class types.

We automatically define operators for Enum class (only if strongly typed)
types, just include this header and the magic will happen now.

Internally using a wrapper struct to wrap the return values so that we
can define the bool() operator and so no need to use the !! operator
anymore for simple true checks.