jsapi-util: Make SFINAE on constructor to work with clang++
As per C++ standard is perfectly legal for the compiler to check
template definitions both during instantation (so looking at potential
issues) and after their actual usage, so while g++ is more liberal on this
and will not try to define functions which are disabled when allocating
the template, clang++ will do it.

So, in the case of using a void pointer to hold GjsAutoPointer it will
complain when using void[] type.

To avoid this, let's change the array constructor definition to use an
implicit template value, while we still check it properly as part of the
enable_if check, so that we'll eventually have the same result.

Also fix other compilation issues due to clang being a bit more
stricter.