function: Compute the array length coherently to what we do in cache
When computing an array length to pass to a function call we bother
some JSapi code to get a value (that also may be of an incorrect type)
while we already compute this value into the arg-cache depending on the
GIArgument type.

Since this is the correct way of doing this and avoids lots of unneeded
checks, let's just expose the function and reuse it.

This is also something that will be quite useful to support BigInt as in
such case we may create objects instead of numbers, causing this to
fail.