value: Support converting BigInt to (u)int64 values
BigInt values are now accepted as parameters to C functions that take
64-bit integers.

For now BigInts are only accepted for 64-bit parameters. We may allow
BigInt for any parameter in the future (once we've checked that we don't
go out-of-range), for now we assume that BigInt values are used only for
functions taking a type that can mostly handle them.

For bounds checks, given that we don't have a bigger value holder than
(u)int64_t, we have to create a new wrapper to avoid taking the fast path
in gjs_arg_set_from_js_value(), and without having to hack
type_has_js_getter().

As per this, now for 64bit integers we use TypeWrapper as the type holder
that has conversion operators to the base type, and for which we can then
add a js_value_to_c_checked() specialization.

See: #271