arg-cache: Support passing GValue objects, not only generated-gvalues
When we create a GValue using the actual constructor we're holding
internally a boxed type, and in such case we don't have to try to
convert it to another GValue to pass around, but we should just use it
as our argument.

However, this can't be computed when parsing the argument, but depends
case-by-case, as per this when we've a GValue and this is an object
containing an actual GValue, pass it around and remember that we've not
to free the returned argument, as that will be handled by the boxed
wrapper.

This allows to finally use some Gtk functions to get GValue parameters
as "in/out" parameters, and the gvalue copy function.
Add tests checking that this is the case.

Fixes #74