arg-cache: Rewrite it taking advantage of C++ features
Currently for each function we have an argument cache that that for each
function initialize an array of equal structures that might contain
information that aren't needed for the type we're handling, thus some
waste of space. Then we use other structures to handle the marshallers.

C++ allows us however to have virtual functions though (which eventually
take the same space that our marshallers pointer would), and also
supports multi-inheritance and using those two things we can make the
argument cache a bit smarter and use more C++ code.

As per this, added an ArgsCache struct that we use inside Function to
manage the arguments without having to care about the internal
implementation at that level.

Then we define a generic virtual Argument struct and we implement
internally using different types depending on the marshalling action we
want to do.

As per this, we can convert most of our functions to C++ functions
quite easily, and leave the memory management to the smart pointers.