arg-cache: Create Return and Instance arguments only if we have one
This is something that also happened in the original cache implementation,
but only for the instance parameter.

Now, given that we handle differently the "skip-all" (AKA void) return
value, we can also avoid to allocate memory for it, if there is none.

This saves some more space compared to the previous iteration on the
cache itself, but also make things safer as the access to the elements
is protected from the outside and we use a constexpr function to get the
actual element index, without having to guess much.