function: Don't use alloca for CallState values
Usage of alloca() is at least discouraged, especially in a project
that it is supposed to work in multiple architectures and OSes, as its
implementation is very hw/compiler dependent.

Not only, when a function has many arguments we may end up in filling up
the stack space, leading to unexpected behavior.

It also prevents us from usage of C++ constructors and destructors in
GjsFunctionCallState.

As per this, remove g_newa() and rely on default C++ allocators, while
use some utilities to compute the proper offsets.