arg-cache: Save GType information for RegisteredType
This is a reimplementation of commit d526bf8, but re-applied on top of
this branch, to make it clearer. The rationale stays:

As found in https://gitlab.gnome.org/GNOME/gjs/-/issues/361, getting the
GType using g_registered_type_info_get_g_type() is very expensive and
can make up for a significant part of the overhead when calling into a C
function.

Since the argument cache seems to be fairly small in a typical
gnome-shell session (about 1300 entries), the total increased memory
usage of about 10kB seems very reasonable given the benefits of the
caching.