arg-cache: Only store GType and InfoType in RegisteredType's
On gjs a registered type can be marshalled by just using its gtype
without strict need of the base info (that it's used mostly for debug
infos), so we can avoid storing such pointer and the related base info
data.

To achieve this we still have to handle the case in which we've an
unregistered Boxed type (so with no GType), and we can do it by adding
a new BoxedIn type that also inherits from BaseInfo and that implements
the newly added boxed info() virtual method.

As per this the actual size of all the Boxed types lower to 40 bits,
without having to keep any base info allocated.