arg-cache: Add marshallers for C arrays of basic types
When a C array uses basic types for contents, we can just keep in cache a
few bits of information about the array type and the element tag.

So, override ExplicitArray and implement its marshallers by inheriting
from defaults.

(Rewritten to not synthesize a GITypeInfo from the GITypeTag by Philip.)