arg-cache: Use a templated structure to generate arguments for containers
Basic type containers arguments can be easily marshalled reusing always
the same pattern, so just use a simple templated structure to generate
the actual code, to which we pass the base marshaller class and the
specific type structure.

Then a constexpr if's based function can be used to determine what to do
in the marshalling functions.

(Refactored to match the other refactors in this branch by Philip.)