arg: Pass GjsArgumentFlags to all the functions and handle IN + NOTHING
In arguments we introduced the argument flags to pass parameters to the
argument parsing functions, however to handle the "in" argument release
we used a workaround extending GITransfer for the specific transfer
none when used with in arguments we allocated.

In order to allow GjsArgumentFlags to provide this information, and not
have to abuse GITransfer anymore, add IN/OUT explicit flags for arguments.
While these could be defined as inverse flags using SKIP_IN, SKIP_OUT,
it's better to have explicit values so that we can pass them to the arg
parsing functions when we need to do specific operations for IN/OUT args.