arg: Keep track of allocated inout arguments
When a function uses inout arguments with transfer container, we are the
one responsible for freeing the allocated container values, while we
assume that the called function freed the container itself.

Now, in such case unfortunately given that we can't assume that the
container pointer is valid once the function is called (and actually
should not, as per the introspection data), the only safe way we have is
to track information about the container and a list of allocated pointers.

So, add a simple caching support for such containers, that allows to
track the container type and for each value a link with the parent
container. This allows to both manually free the container elements and
to have a free function to be used as a DestroyNotify callback for the
types supporting it, being the safest solution to also support callers
freeing or destroying the container.

In fact, while for some types (GArray, GPtrArray or GHash) we could use a
more optimized path, by just adding an extra reference to such containers,
unfortunately we can't be sure about the safety of this as they also
provide _free() and _destroy() functions that would break our
assumption.

Fixes: #44