jsapi-util: Add a function to check if a BigInt is out of range
BigInts can contain (u)int64_t but they also can hold values that are way
bigger or smaller than their limits, so in order to ensure that we can do
runtime checks of the C-representability of a BigInt we need to verify
that it stays within the type boundaries we care about.

Since version 91 mozjs provides a C++ API for this, JS::BigIntFits(). We
use this function instead of calling JS::BigIntFits() directly, so that we
can additionally clamp the number to the type's limits if it doesn't fit.

See: #271