overrides/Gtk: Set BuilderScope in class init
The API to set a custom template scope is on the class, not the instance.

On Gtk3 we don't have an alternative for hooking up the correct "this"
argument (as far as I can tell), but we do have the option with Gtk4.

Moving the code is not only more correct, it is particularly important
as Gtk4 now relies entirely on ref-counting for destroying widgets
(and thus ultimately for exiting a Gtk.Application).

Currently the separate reference to the "this" object in the custom
scope prevents the ref-count of any template widget from ever dropping
to 0, keeping the app running indefinitely after closing all windows.