gi/arg-cache: Only skip array length parameter once
If an array-length parameter precedes the corresponding array parameter,
we skip counting the latter instead of the former. As the parameter type
isn't relevant for the argument counter, that works fine - provided that
the length parameter is only used for a single array.

That is usually the case, but there are exceptions (for example Gtk4's
gtk_accessible_update_property_value()). Address this by only skipping
array arguments if the corresponding length parameter wasn't already
marked skip-all previously.