CI: Add cppcheck config file
This is needed because we are about to upgrade to a newer version of
cppcheck than the one on the DockerHub image. It ships a "library file"
for the GNOME platform, enabling some extra checks when using GLib and GTK
APIs. It also requires configuration of some of the macros that we use.