CI: Fix bug in IWYU mapping file
Oops, we accidentally made the IWYU mapping file invalid, so the IWYU job
was not functioning correctly. Move the workaround into the postprocessing
script, and fix all the include fixes that have been missed in the
meantime while this was broken.

New code has also exposed some false positives in IWYU, so add those
exceptions to the false positive list.