js: Use strings and stringstreams instead of GString
std::ostringstream is the C++ standard library version of GString. If we
are not doing any formatting but only appending other strings, then we can
even just use std::string and the + operator.

This allows using RAII so we can get rid of a goto, and just generally be
more memory-safe.

We continue using GString in gjs_hyphen_from_camel() for performance
reasons, because we would like a modifiable C buffer that we can pass to
canonicalize_key(), and the only way to get a modifiable buffer out of
std::string would be to copy it.

Note: stream.str() _creates_ a std::string object, and copies the string
data. string.c_str() returns a const char* _without_ copying.