debugger: Make '$$' mean the last value
The debugger already stores the result of each printed expression in
variables named $1, $2, etc., so you can refer to them later. This change
makes the variable '$$' always refer to the last printed value.

Adds a bit of documentation explaining how these variables can be printed,
and adds a test for $$ as well as $1, $2, etc.

Based on a commit from Mozilla's original jorendb:
https://hg.mozilla.org/mozilla-central/rev/fe52996292bb6d1acd51a983f9a7ac6faa7724f6
(The GJS debugger was originally based on jorendb and I was looking to see
if any new improvements had been made since I had forked it)