test: Fix debugger test script to print diff properly
We added this test for the exit code because the tests weren't failing if
one of the commands earlier in the pipeline failed, but mistakenly
suppressed printing a diff of the output (because diff always "fails" if
there is a difference.)

If there is a diff, print it; only if there is no diff, look at the exit
code.

Unreviewed, pushing to make developing the debugger easier.