CI: Compare cpplint and IWYU against the pull request's base branch
Instead of always comparing against GNOME/gjs/master, try to find the
place where the current branch was branched off of the upstream repository
using an adaptation of the script that GLib uses.

The main adaptation compared to GLib is that we are running on a
lightweight image with sh and busybox, so we don't have fancy things like
process redirection and diff formatting options.