CI: Enable full cppcheck
Enable a full scan of all of the C++ files with cppcheck. We don't have
any build dependencies on the Alpine Linux image so move the cppcheck job
to a later stage where we can access a compile commands database and
generated C files as an artifact of an earlier job.