gtype: Remove JSClass macros
Port gtype.cpp to use the CWrapper template, instead of the
GJS_DEFINE_PRIV_FROM_JS family of macros. This is slightly different than
Ns because GTypeObj is a class with only static methods, whereas the
actual data stored in the object's private pointer slot is the GType,
stuffed into a void*.

Also requires adding a debug topic for GTypes.