js: Add new system for creating wrapper objects for native types
In order to phase out the GJS_DEFINE_PRIV_FROM_JS set of macros in
jsapi-class.h, we introduce a new template class, CWrapper. This is like
the GIWrapper family of classes in gi/wrapperutils.h, but without any
usage of gobject-introspection. CWrapper simply implements a JS object
that wraps a C pointer and stores its prototype in a private slot on the
global object. It will be used for GI namespace objects, GI function
objects, legacy importer objects, GType objects, and Cairo objects.

CWrapper does most of the things that the macros did already, but with
more type safety.

In CWrapper we avoid using JS_InitClass() which the SpiderMonkey
developers consider broken for various reasons. We use js::ClassSpec
instead. The main change is that we now avoid the weird behaviour of
JS_InitClass() where the constructor is always defined as a global object
and the even weirder behaviour that if there is no constructor, then the
prototype is defined as a global object. I don't believe this happened
every time anyway, since there wasn't always a global object passed in,
and I don't believe this will break any code in practice, but if it does,
we can also add that behaviour to this template class.

In addition, functionality (for_js(), for_js_nocheck(), for_js_typecheck()
and typecheck()) that is common to both CWrapper and GIWrapperBase gets
moved into a separate CRTP class, CWrapperPointerOps. We do have to give
GIWrapperBase::for_js_typecheck() a slightly different signature than it
previously had, because in this more general method we need to distinguish
a null pointer from the error state.