wrapperutils: Really make debug methods no-ops if compiled out
The debug methods are supposed to be no-ops if verbose logging is not
compiled in, but what was happening was that the jsid and JSString were
being converted to UTF-8 because only the code inside gjs_debug_jsprop()
is ifdef'd out if not built with verbose logging. So this was potentially
slowing down performance.

Instead use constexpr to make these calls only if verbose logging is
compiled in.