CI: Fix getting the upstream base
Use git-merge-base to calculate the newest common ancestor, which should
be more reliable.

Also increase the horizon to 28 days since the newest common ancestor of
this very merge request is more than 14 days old. Check that any commits
were fetched (if not, "the remote end hung up unexpectedly") and fetch the
last 30 commits instead if it didn't work. (Code review suggestion from
Evan)

As well, it seems that this feature does not always work quite correctly,
so it would be good to have some debug output.