debugger: Handle special return values from Environment.getVariable()
Debugger.Environment.getVariable() may return one of the three sentinel
values {missingArguments: true}, {optimizedOut: true}, or {uninitialized:
true} if a variable does not have a value for some reason. Handle these
special values in debuggeeValueToString().

"missingArguments" means that a function argument was optimized out,
"optimizedOut" means that a regular variable was optimized out, and
"uninitialized" means that a variable is present in the environment but
hasn't been initialized yet, for example when breaking on an exception
thrown while initializing it.

Also adds a test for the latter case (printing a backtrace with locals,
when breaking on an exception that leaves a local variable uninitialized.)

https://searchfox.org/mozilla-central/rev/6a6a366031680829746b5d2362610b868fd9571a/js/src/debugger/Debugger.cpp#1412-1435