arg: fix build failure with glib master
GLib has changed the definition of glib_typeof in glib!1715. I don't
fully understand why, but that has broken gjs here. The good news is I
don't need to understand: here we are using g_steal_pointer on a smart
pointer that has a release function. It's nicer to just use release
instead and sidestep the issue.