esm: Add additional builtin ES modules
Adds 'cairo', 'gettext', and 'system' modules, to match the builtin legacy
modules that GJS already had.

'gettext' and 'system' have named exports because it's expected that the
functions might be used individually that way. That doesn't make as much
sense for Cairo which is to be understood more as if it were a GI module.

(changes from Philip folded in: Jasmine tests)