GObject: Guard against classes with GType name 'Gjs_'
Previously, if an anonymous class expression was passed to
GObject.registerClass(), then the GType name would become 'Gjs_', meaning
that only one of these classes could ever be registered in the same
process.

Instead, use a name based on a UUID when the class expression is
anonymous. I chose a UUID, and not a sequence number, for example, so that
it is random and cannot be the same in any two given runs.