jsapi-util-string: Improve gjs_debug_object() to print more info
Previously we'd print the name of the function if dumping a function
object that was in a JS::Value but not if dumping the object directly.
Move that code into gjs_debug_object(), and also add more code to dump
information about a Promise object.

Like Symbols, Promise objects don't require a JSContext to get information
about them and can't trigger a GC, so they are safe to debug-dump here.