README.MSVC.md: Mention about VS2019 requirement
Update README.MSVC.md to mention that in order to build with Visual
Studio, the 2019 version is now required.

Currently, with the fixes in this MR, GJS master will build on Visual
Studio 2019 as well as clang-cl, but one must use `/permissive` instead
of the default `/permissive-` for the time being before #368 is
resolved.