console.cpp: Make code build on Windows
The new code that tries to detect Ctrl+C almost built fine on Windows,
except that Windows does not have sigjmp_buf, siglongjmp() and
sigsetjmp().  For Windows we have near-equivilents jmp_buf, longjmp()
and setjmp(), so use them.  The mask parameter for sigsetgmp() does
not apply on Windows, so just discard it when calling setjmp() in its
place.