enum-utils.h: Fix type ambiguity on Visual Studio
Make the wrapper use the underlying type on Visual Studio builds, so
that we can pinpoint on the "right" bitwise operator to use on Visual
Studio builds, for the class enumeration(s) we may use.

Also, on the same token, define casts for Visual Studio for the wrapped
types for use on the '|=' and '&=' operators, as well as the '&' and '|'
operators that will be used in value assignments so that the compiler
does not get upset when converting between the underlying type and the
class enumeration type.