js-value-inl: Consider NaN results to be out of range values for integers
When converting to integers using a bigger container (such as double) we
may end up having our value to be set to NaN, which JS::ToNumber would
return in case the JS value is undefined.

In such case we should avoid to silently use the NaN value, but instead
mark it as an out of range value as well.