jsapi-util-string: Check for overflow when copying string chars
A previously existing bug was that calculating the number of bytes to copy
might overflow if the number of JS characters was not over the max of
size_t but the number of bytes was. Check for this unlikely case (a string
of 2.1 billion characters on 32-bit, and an astronomical number on 64-bit)
and bail.

Thanks to Bastien Nocera for the suggestion.