function: Save original allocated pointers
cppcheck is not smart enough to determine that when we save an allocated
pointer plus an offset, and then free the saved pointer minus that offset,
it amounts to the same thing. But, maybe it is clearer anyway to save the
original pointers and add the offset in an accessor.