object: Do not call any function on disposed GObject pointers
After an object is disposed we should really not perform anything with
it, however we may end up calling some GObject functions during
disposition.

In particular, we try to print its name, unset the qdata and unref it
again.

So during disassociation always ensure that the object we're
disassociating has not been already destroyed, before doing such
operations.