CI: Skip refcount tests in extra_gc mode
extra_gc mode causes an extra GC after every allocation. This causes the
refcount tests to fail because the JS reference to the object is dropped
since nothing is holding on to it from user code. However, we cannot force
this extra GC to happen from user code, since it occurs in the middle of
C++ marshalling. Therefore, we have to skip these tests in extra_gc mode.