context: Improve profiling data for garbage collections
This sends all the timing information to the profiler, that we get from
the JS engine's GC and finalization status callbacks. Instead of just the
sweeping timings, we keep track of the duration of the whole GC as well as
the sweeping of zone groups, and we report the reason for the GC that the
engine gives. For cases where GJS triggers the GC, we add a few more
GC reason constants with explanations that will show up in Sysprof.