system: Add System.dumpMemoryInfo() API
SpiderMonkey provides a "memory info object" with a bunch of slightly
obscure GC statistics on it. We don't want to expose this object directly
to JS code, but we can dump it to a file or stdout, as we do with
System.dumpHeap().

This adds a System.dumpMemoryInfo() API that implements this.

We also take the two most pertinent statistics (bytes allocated in
garbage-collectable objects and malloc bytes associated with them) and
pipe these to Sysprof when System.dumpMemoryInfo() is called.

We can't query these statistics in the SIGPROF handler because doing so
would allocate memory. So if you want a continuous display of these
statistics in Sysprof, then you need to call System.dumpMemoryInfo() in a
timeout callback or something like that.

See: #292